+++ mod/app/src/test/java/bttv/mod/RetokenizeTest.java
@@ -15,6 +15,7 @@ import java.util.Arrays;
 import java.util.Collection;
 
 import bttv.Data;
+import bttv.Res;
 import bttv.Tokenizer;
 import bttv.emote.Emote;
 import bttv.emote.Emotes;
@@ -199,6 +200,8 @@ public class RetokenizeTest {
         when(mockContext.getSharedPreferences("BTTV", 0))
                 .thenReturn(mockPrefs);
         when(mockPrefs.getBoolean("enable_ffz_emotes", true)).thenReturn(true);
+        when(mockPrefs.getString("bttv_gif_render_mode", Res.strings.bttv_settings_gif_render_mode_animate.name()))
+                .thenReturn(Res.strings.bttv_settings_gif_render_mode_animate.name());
         Data.setContext(mockContext);
         Data.setCurrentBroadcasterId(10);
         Emotes.addChannelFFZ(10,