+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 # master
 - update base to 11.1.0
+- add static versions of Gif Emotes
+- fix crash when opening friends page
+- fix broken strings
 
 # v0.8.1-beta
 - adds highlight keywords feature